Removed 'check' script function, altered LOAD DATA LOCAL routine
Before, the loadcalmap handled the 'check' script and all, but now
we have separated these two, and replaced 'check' with 'update', which
is run separately.

Also, there was a bug regarding the location of the file that loadcalmap
is handed. This has been taken care of, partly due to MySQL's knowledge
of where your present working directory is.